consolidate get universe information, add open interest to universe object.
