+++ src/modules/account/components/account-universe-description/account-universe-description.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Styles from 'modules/account/components/account-universe-description/account-universe-description.styles'
-import { formatAttoRep } from 'utils/format-number'
+import { formatAttoRep, formatAttoEth } from 'utils/format-number'
 
 const AccountUniverseDescription = p => (
   <div className={Styles.AccountUniverseDescription}>
@@ -25,6 +25,7 @@ const AccountUniverseDescription = p => (
     <div className={Styles.AccountUniverseDescription__description}>
       <div>Your wallet:  {p.accountRep === '0' ? '0' : formatAttoRep(p.accountRep, { decimals: 2, roundUp: true }).formatted} REP</div>
       <div>Total amount of REP:  {p.accountRep === '0' ? '0' : formatAttoRep(p.universeRep, { decimals: 2, roundUp: true }).formatted} REP</div>
+      <div>Total Open Interest:  {p.openInterest === '0' ? '0' : formatAttoEth(p.openInterest, { decimals: 2, roundUp: true }).formatted} ETH</div>
       <div>Number of markets created:  {p.numMarkets}</div>
     </div>
   </div>
@@ -33,6 +34,7 @@ const AccountUniverseDescription = p => (
 AccountUniverseDescription.propTypes = {
   switchUniverse: PropTypes.func.isRequired,
   accountRep: PropTypes.string.isRequired,
+  openInterest: PropTypes.string.isRequired,
   universeRep: PropTypes.string.isRequired,
   numMarkets: PropTypes.number.isRequired,
   isWinningUniverse: PropTypes.bool.isRequired,
